ku-img-backend/app/autotag/img/crawlers/requirements.txt
ImageHash
Pillow
Requests
selenium
webdriver_manager
